+++ R/prepare_ma.R
@@ -10,12 +10,14 @@
 #' @param effect what effect to calculate? a `mean` (and SE) of outcome in groups or
 #'               (for binary data) `logOR` (odds ratio), `logRR` (risk ratio);
 #' @param log logical; log-transform the outcome variable?
-#' @param rare_event_correction If effect is `logOR` or `logRR`, this correction
-#'             is used when working with
-#'             binary data only. The value of correction is added to all arms
-#'             in trials where some arms had 0 events.
+#' @param rare_event_correction This correction is used when working with
+#'             binary data (effect `logOR` or `logRR`)
+#'             The value of correction is added to all cells in
+#'             either some or all rows (groups), depending on `correction_type`.
 #'             Using corrections may bias results but is the only alternative to
 #'             avoid infinite values.
+#' @param correction_type If `"single"` then rare event correction is only applied to the
+#'             particular rows that have 0 cells, if `"all"`, then to all studies
 #' @param cfb logical; calculate change from baseline? If yes, the outcome
 #'            variable is taken as a difference between values in `outcome` and
 #'            `baseline` columns
@@ -48,21 +50,40 @@
 #'
 #' @author Witold Wiecek
 #' @seealso [convert_inputs] for how any type of data is (internally) converted into
-#'          a list of Stan inputs;
+#'          a list of Stan inputs; vignette `baggr_binary` for more details about
+#'          rare event corrections
 #' @export
 #' @import stats
 #'
+#' @examples
+#'
+#' # Example of working with binary outcomes data
+#' # Make up some individual-level data first:
+#' df_rare <- data.frame(group = paste("Study", LETTERS[1:5]),
+#'                       a = c(0, 2, 1, 3, 1), c = c(2, 2, 3, 3, 5),
+#'                       n1i = c(120, 300, 110, 250, 95),
+#'                       n2i = c(120, 300, 110, 250, 95))
+#' df_rare_ind <- binary_to_individual(df_rare)
+#' # Calculate ORs; default rare event correction will be applied
+#' prepare_ma(df_rare_ind, effect = "logOR")
+#' # Add 0.5 to all rows
+#' prepare_ma(df_rare_ind, effect = "logOR",
+#'            correction_type = "all",
+#'            rare_event_correction = 0.5)
 
 prepare_ma <- function(data, #standardise = NULL,
                        effect = c("mean", "logOR", "logRR"),
                        rare_event_correction = 0.25,
+                       correction_type = c("single", "all"),
                        log = FALSE, cfb = FALSE, summarise = TRUE,
                        treatment="treatment",
                        baseline = NULL,
                        group="group",
                        outcome="outcome") {
 
-  effect <- match.arg(effect, c("mean", "logOR", "logRR"))
+  effect <- match.arg(effect)
+  correction_type <- match.arg(correction_type)
+
   if(grepl("pool|unknown", detect_input_type(data, group, treatment, outcome))){
     if(effect %in% c("logOR", "logRR")){
       check_columns_binary(data)
@@ -195,16 +216,25 @@ prepare_ma <- function(data, #standardise = NULL,
                  b     = sum(treatment == 1) - sum(outcome[treatment == 1]),
                  d     = sum(treatment == 0) - sum(outcome[treatment == 0])))
         }))
-      rare <- with(binary_data_table, (a == 0 | b == 0 | c == 0 | d == 0))
-      if(sum(rare) > 0 && rare_event_correction == 0.25)
-        message("Applied default rare event correction (0.25) in ", sum(rare), " studies")
 
-      binary_data_table$a  <- v*rare + binary_data_table$a
-      binary_data_table$b  <- v*rare + binary_data_table$b
-      binary_data_table$c  <- v*rare + binary_data_table$c
-      binary_data_table$d  <- v*rare + binary_data_table$d
-      binary_data_table$n1 <- v*rare + binary_data_table$n1
-      binary_data_table$n2 <- v*rare + binary_data_table$n2
+      rare <- with(binary_data_table, (a == 0 | b == 0 | c == 0 | d == 0))
+
+      if(sum(rare) == 1 && rare_event_correction == 0.25)
+        message("Applied default rare event correction (0.25) in 1 study")
+      if(sum(rare) > 1 && rare_event_correction == 0.25)
+        message("Applied default rare event corrections (0.25) in ", sum(rare), " studies")
+
+      if(correction_type == "single")
+        cc_value <- v*rare
+      if(correction_type == "all")
+        cc_value <- rare_event_correction
+
+      binary_data_table$a  <- cc_value + binary_data_table$a
+      binary_data_table$b  <- cc_value + binary_data_table$b
+      binary_data_table$c  <- cc_value + binary_data_table$c
+      binary_data_table$d  <- cc_value + binary_data_table$d
+      binary_data_table$n1 <- cc_value + binary_data_table$n1
+      binary_data_table$n2 <- cc_value + binary_data_table$n2
 
 
       out <- binary_data_table